Updated NavMesh scripts.
