+++ Assets/NavMesh/NavMesh.unity
@@ -134,6 +134,7 @@ GameObject:
   - component: {fileID: 171855654}
   - component: {fileID: 171855653}
   - component: {fileID: 171855652}
+  - component: {fileID: 171855655}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -207,6 +208,20 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 55.891, y: 0, z: 0}
+--- !u!114 &171855655
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 171855651}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 83cc7339a480343f89d9dbf7ac0357a9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  mainCamera: {fileID: 171855653}
+  agent: {fileID: 651313534}
 --- !u!1 &299055586
 GameObject:
   m_ObjectHideFlags: 0
@@ -351,7 +366,6 @@ GameObject:
   - component: {fileID: 651313532}
   - component: {fileID: 651313531}
   - component: {fileID: 651313530}
-  - component: {fileID: 651313533}
   - component: {fileID: 651313534}
   m_Layer: 0
   m_Name: Capsule
@@ -439,18 +453,6 @@ MeshFilter:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 651313528}
   m_Mesh: {fileID: 10208, guid: 0000000000000000e000000000000000, type: 0}
---- !u!114 &651313533
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 651313528}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 83cc7339a480343f89d9dbf7ac0357a9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!195 &651313534
 NavMeshAgent:
   m_ObjectHideFlags: 0
